+++ src-opam/dockerfile_distro.ml
@@ -19,7 +19,7 @@
 open Astring
 
 type t = [
-  | `Alpine of [ `V3_3 | `V3_4 | `V3_5 | `V3_6 | `V3_7 | `V3_8 | `V3_9 | `V3_10 | `V3_11 | `V3_12 | `Latest ]
+  | `Alpine of [ `V3_3 | `V3_4 | `V3_5 | `V3_6 | `V3_7 | `V3_8 | `V3_9 | `V3_10 | `V3_11 | `V3_12 | `V3_13 | `Latest ]
   | `Archlinux of [ `Latest ]
   | `CentOS of [ `V6 | `V7 | `V8 | `Latest ]
   | `Debian of [ `V10 | `V9 | `V8 | `V7 | `Stable | `Testing | `Unstable ]
@@ -36,7 +36,7 @@ type status = [
 ] [@@deriving sexp]
 
 let distros = [
-  `Alpine `V3_3; `Alpine `V3_4; `Alpine `V3_5; `Alpine `V3_6; `Alpine `V3_7; `Alpine `V3_8; `Alpine `V3_9; `Alpine `V3_10; `Alpine `V3_11; `Alpine `V3_12; `Alpine `Latest;
+  `Alpine `V3_3; `Alpine `V3_4; `Alpine `V3_5; `Alpine `V3_6; `Alpine `V3_7; `Alpine `V3_8; `Alpine `V3_9; `Alpine `V3_10; `Alpine `V3_11; `Alpine `V3_12; `Alpine `V3_13; `Alpine `Latest;
   `Archlinux `Latest;
   `CentOS `V6; `CentOS `V7; `CentOS `V8; `CentOS `Latest;
   `Debian `V10; `Debian `V9; `Debian `V8; `Debian `V7;
@@ -50,8 +50,9 @@ let distros = [
 
 let distro_status (d:t) : status = match d with
   | `Alpine (`V3_3 | `V3_4 | `V3_5 | `V3_6 | `V3_7 | `V3_8 | `V3_9 | `V3_10 | `V3_11) -> `Deprecated
-  | `Alpine `V3_12 -> `Active `Tier1
-  | `Alpine `Latest -> `Alias (`Alpine `V3_12)
+  | `Alpine `V3_12 -> `Active `Tier3
+  | `Alpine `V3_13 -> `Active `Tier1
+  | `Alpine `Latest -> `Alias (`Alpine `V3_13)
   | `Archlinux `Latest -> `Active `Tier3
   | `CentOS `V8 -> `Active `Tier2
   | `CentOS `V7 -> `Active `Tier3
@@ -96,7 +97,7 @@ let distro_arches ov (d:t) =
   match resolve_alias d, ov with
   | `Debian `V10, ov when OV.(compare Releases.v4_05_0 ov) = -1 -> [ `I386; `X86_64; `Aarch64; `Ppc64le; `Aarch32 ]
   | `Debian `V9, ov when OV.(compare Releases.v4_05_0 ov) = -1 -> [ `I386; `X86_64; `Aarch64; `Aarch32 ]
-  | `Alpine (`V3_6 | `V3_7 | `V3_8 | `V3_9 | `V3_10 | `V3_11 | `V3_12), ov when OV.(compare Releases.v4_05_0 ov) = -1 -> [ `X86_64; `Aarch64 ]
+  | `Alpine (`V3_6 | `V3_7 | `V3_8 | `V3_9 | `V3_10 | `V3_11 | `V3_12 | `V3_13), ov when OV.(compare Releases.v4_05_0 ov) = -1 -> [ `X86_64; `Aarch64 ]
   | `Ubuntu (`V18_04|`V20_04|`V20_10), ov when OV.(compare Releases.v4_05_0 ov) = -1  -> [ `X86_64; `Aarch64; `Ppc64le ]
   | `Fedora (`V33), ov when OV.(compare Releases.v4_08_0 ov) = -1  -> [ `X86_64; `Aarch64 ]
   | _ -> [ `X86_64 ]
@@ -152,6 +153,7 @@ let builtin_ocaml_of_distro (d:t) : string option =
   |`Alpine `V3_10 -> Some "4.07.0"
   |`Alpine `V3_11 -> Some "4.08.1"
   |`Alpine `V3_12 -> Some "4.08.1"
+  |`Alpine `V3_13 -> Some "4.08.1"
   |`Archlinux `Latest -> Some "4.11.1"
   |`Fedora `V21 -> Some "4.01.0"
   |`Fedora `V22 -> Some "4.02.0"
@@ -237,6 +239,7 @@ let tag_of_distro (d:t) = match d with
   |`Alpine `V3_10 -> "alpine-3.10"
   |`Alpine `V3_11 -> "alpine-3.11"
   |`Alpine `V3_12 -> "alpine-3.12"
+  |`Alpine `V3_13 -> "alpine-3.13"
   |`Alpine `Latest -> "alpine"
   |`Archlinux `Latest -> "archlinux"
   |`OpenSUSE `V42_1 -> "opensuse-42.1"
@@ -301,6 +304,7 @@ let distro_of_tag x : t option = match x with
   |"alpine-3.10" -> Some (`Alpine `V3_10)
   |"alpine-3.11" -> Some (`Alpine `V3_11)
   |"alpine-3.12" -> Some (`Alpine `V3_12)
+  |"alpine-3.13" -> Some (`Alpine `V3_13)
   |"alpine" -> Some (`Alpine `Latest)
   |"archlinux" -> Some (`Archlinux `Latest)
   |"opensuse-42.1" -> Some (`OpenSUSE `V42_1)
@@ -364,6 +368,7 @@ let rec human_readable_string_of_distro (d:t) =
   |`Alpine `V3_10 -> "Alpine 3.10"
   |`Alpine `V3_11 -> "Alpine 3.11"
   |`Alpine `V3_12 -> "Alpine 3.12"
+  |`Alpine `V3_13 -> "Alpine 3.13"
   |`Archlinux `Latest -> "Archlinux"
   |`OpenSUSE `V42_1 -> "OpenSUSE 42.1"
   |`OpenSUSE `V42_2 -> "OpenSUSE 42.2"
@@ -425,6 +430,7 @@ let base_distro_tag ?(arch=`X86_64) d =
         | `V3_10 -> "3.10"
         | `V3_11 -> "3.11"
         | `V3_12 -> "3.12"
+        | `V3_13 -> "3.13"
         | `Latest -> assert false
       in
       match arch with